more efficient and correct sequences

+++ include/Sequence.h
@@ -1,19 +1,21 @@
 #ifndef SEQUENCE_H
 #define SEQUENCE_H
 
+#include <cstdlib>  // size_t
+
 template <class T> class Sequence {
 public:
     virtual ~Sequence() {}
 
     virtual T GetFirst() const = 0;
     virtual T GetLast() const = 0;
-    virtual T Get(int index) const = 0;
-    virtual Sequence<T>* GetSubsequence(int startIndex, int endIndex) const = 0;
-    virtual int GetLength() const = 0;
+    virtual T Get(size_t index) const = 0;
+    virtual Sequence<T>* GetSubsequence(size_t startIndex, size_t endIndex) const = 0;
+    virtual size_t GetLength() const = 0;
 
     virtual Sequence<T>* Append(T item) = 0;
     virtual Sequence<T>* Prepend(T item) = 0;
-    virtual Sequence<T>* InsertAt(T item, int index) = 0;
+    virtual Sequence<T>* InsertAt(T item, size_t index) = 0;
     virtual Sequence<T>* Concat(Sequence<T>* list) = 0;
 
     virtual Sequence<T>* Map(T (*func)(T)) const = 0;
